перемещены повторяющие проверки в Before и After, удален класс Negative, тесты из него перенесены в существубщие
